Improve Concourse (#70)
* Concourse 3.2 Syntax

This change updates the pipeline definition to use the proper substitution
syntax for Concourse 3.2 and later.

* YAML Reference

This change updates the pipeline to make use of YAML references.  This allows
common blocks of code (like parameter groups) to be defined once and shared
between jobs.

* Formatting Consistency

This change modifies the format a bit to make it consistent across the entire
file.  This mostly has to do with the indention of YAML arrays.

* Improve Generate Version

The change improves the generate-version task. It does this by using a more
descriptive output name, using a local clone as recommended by the Concourse
team, and only configuring the local git repository.