Reduce duplication
It would make sense to go a step further and move the message that's
printed to be done from within the script that's being called.

It would also be appropriate to create a new task for tagging the
repository; build-and-upload and prod-deploy are the only tasks that do
this and are otherwise identical to the other tasks.